requirements.txt
# Core framework
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8
django-allauth==65.4.1

# Environment management
environs==14.1.1

# HTTP client
httpx==0.28.1

# LLM providers
groq==0.24.0
litellm==1.67.6
langchain==0.3.27


pymupdf==1.26.3